+++ apps/endringsmelding-pleiepenger/src/app/components/buttons/EditButton.tsx
@@ -3,15 +3,7 @@ import React from 'react';
 import { Edit } from '@navikt/ds-icons';
 
 const EditButton: React.FunctionComponent<ButtonProps> = ({ ...rest }) => {
-    return (
-        <Button
-            icon={<Edit role="presentation" aria-label={`Endre`} />}
-            type="button"
-            variant="primary"
-            size="small"
-            {...rest}
-        />
-    );
+    return <Button icon={<Edit role="presentation" />} type="button" variant="primary" size="small" {...rest} />;
 };
 
 export default EditButton;