+++ apps/endringsmelding-pleiepenger/src/app/utils/initialDataUtils.ts
@@ -3,11 +3,11 @@ import dayjs from 'dayjs';
 
 export const getPeriodeForArbeidsgiverOppslag = (
     dateRangeAlleSaker: DateRange,
-    tillattEndringsperiode: DateRange
+    tillattEndringsperiode: DateRange,
 ): DateRange | undefined => {
     const dateRange = {
-        from: dayjs.max(dayjs(dateRangeAlleSaker.from), dayjs(tillattEndringsperiode.from)).toDate(),
-        to: dayjs.min(dayjs(dateRangeAlleSaker.to), dayjs(tillattEndringsperiode.to)).toDate(),
+        from: dayjs.max(dayjs(dateRangeAlleSaker.from), dayjs(tillattEndringsperiode.from))!.toDate(),
+        to: dayjs.min(dayjs(dateRangeAlleSaker.to), dayjs(tillattEndringsperiode.to))!.toDate(),
     };
     if (dayjs(dateRange.to).isBefore(dateRange.from)) {
         return undefined;