+++ packages/sif-common-formik-ds/src/components/formik-date-range-picker/dateRangePickerUtils.ts
@@ -30,18 +30,18 @@ export const getRangesEndingBeforeDate = (date: Date, dateRanges: DateRange[] = 
 
 export const findClosestDateAfterDate = (fromDate: Date, dates: Date[]): Date | undefined => {
     const dayjss = dates.map((d) => dayjs(d)).filter((m) => m.isAfter(fromDate, 'day'));
-    return dayjss.length > 0 ? dayjs.min(dayjss).toDate() : undefined;
+    return dayjss.length > 0 ? dayjs.min(dayjss)!.toDate() : undefined;
 };
 
 export const findClosestDateBeforeDate = (fromDate: Date, dates: Date[]): Date | undefined => {
     const dayjss = dates.map((d) => dayjs(d)).filter((m) => m.isBefore(fromDate, 'day'));
-    const maxDate = dayjss.length > 0 ? dayjs.max(dayjss).toDate() : undefined;
+    const maxDate = dayjss.length > 0 ? dayjs.max(dayjss)!.toDate() : undefined;
     return maxDate;
 };
 
 export const findClosestDateBeforeOrEqualDate = (fromDate: Date, dates: Date[]): Date | undefined => {
     const dayjss = dates.map((d) => dayjs(d)).filter((m) => m.isSameOrBefore(fromDate, 'day'));
-    const maxDate = dayjss.length > 0 ? dayjs.max(dayjss).toDate() : undefined;
+    const maxDate = dayjss.length > 0 ? dayjs.max(dayjss)!.toDate() : undefined;
     return maxDate;
 };
 
@@ -57,7 +57,7 @@ export const getClosestDateRangeAfterDate = (date: Date, ranges: DateRange[]): D
 
 const getFollowingDate = (
     dateRange: DateRange | undefined,
-    allowRangesToStartAndStopOnSameDate?: boolean
+    allowRangesToStartAndStopOnSameDate?: boolean,
 ): Date | undefined => {
     if (!dateRange) {
         return undefined;
@@ -115,7 +115,7 @@ export const getMaxDateForRangeEnd = ({
 
 const getPreceedingDate = (
     dateRange: DateRange | undefined,
-    allowRangesToStartAndStopOnSameDate?: boolean
+    allowRangesToStartAndStopOnSameDate?: boolean,
 ): Date | undefined => {
     if (!dateRange) {
         return undefined;